requirements.txt
camelot-py==0.9.0
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
cssselect==1.2.0
et-xmlfile==1.1.0
lxml==5.3.0
numpy==2.1.1
opencv-python==4.10.0.84
openpyxl==3.1.5
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.4
pdfquery==0.4.3
pillow==10.4.0
pycparser==2.22
PyPDF2==3.0.1
pypdfium2==4.30.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
pytz==2024.2
roman==4.2
six==1.16.0
tzdata==2024.1
